+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/disabled/testRingAutonPt2.java
@@ -34,6 +34,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.chrisBot;
+import org.firstinspires.ftc.teamcode.ringObject;
 
 import java.util.ArrayList;
